Move exposed module types data structures to can
